Improve styling and remove chip select
Improved formatting, naming, spacing, comments, etc. to improve
readability and comprehension of the code.

Completely removed the chip select from the design since it is not being
used. Additionally, pin 'wea' (Write Enable, Port A) and bus 'dina'
(Data In, Port A) on the character library are assigned values of '0' in
the component declaration since they are not being used; Vivado
throws a warning for each pin (9 total warning) unless they are
declared.